+++ suggest.el
@@ -307,7 +307,10 @@ we look up `t' instead.")
          (< (car args) 0))
     ;; If `read' is called with nil or t, it prompts interactively.
     (and (eq fn 'read)
-         (member args '(nil (nil) (t)))))))
+         (member args '(nil (nil) (t))))
+    ;; Work around https://github.com/magnars/dash.el/issues/241
+    (and (memq fn '(-interleave -zip))
+         (null args)))))
 
 (defface suggest-heading
   '((((class color) (background light)) :foreground "DarkGoldenrod4" :weight bold)